Add bitcoin fragment navigation using a main container
