+++ demo-android/src/main/java/nl/tudelft/ipv8/android/demo/CoinCommunity.kt
@@ -4,8 +4,6 @@ import nl.tudelft.ipv8.Community
 import nl.tudelft.ipv8.IPv8
 import nl.tudelft.ipv8.android.IPv8Android
 import nl.tudelft.ipv8.attestation.trustchain.*
-import nl.tudelft.ipv8.attestation.trustchain.store.TrustChainStore
-import nl.tudelft.ipv8.keyvault.Key
 import java.util.*
 
 class CoinCommunity: Community() {
