+++ demo-android/src/main/java/nl/tudelft/ipv8/android/demo/ui/bitcoin/JoinNetworkFragment.kt
@@ -1,0 +1,39 @@
+package nl.tudelft.ipv8.android.demo.ui.bitcoin
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import nl.tudelft.ipv8.android.demo.R
+import nl.tudelft.ipv8.android.demo.ui.BaseFragment
+
+/**
+ * A simple [Fragment] subclass.
+ * Use the [BitcoinFragment.newInstance] factory method to
+ * create an instance of this fragment.
+ */
+class JoinNetworkFragment : BaseFragment(R.layout.fragment_join_network) {
+    override fun onActivityCreated(savedInstanceState: Bundle?) {
+        super.onActivityCreated(savedInstanceState)
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_join_network, container, false)
+    }
+
+    companion object {
+        /**
+         * Use this factory method to create a new instance of
+         * this fragment using the provided parameters.
+         *
+         * @return A new instance of fragment bitcoinFragment.
+         */
+        @JvmStatic
+        fun newInstance() = JoinNetworkFragment()
+    }
+}
