+++ demo-android/src/main/res/layout/fragment_landing_bitcoin.xml
@@ -2,6 +2,7 @@
 <FrameLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:id="@+id/landing_bitcoin_container"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
@@ -9,18 +10,6 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:orientation="vertical">
-
-        <TextView
-            android:id="@+id/textView2"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="Join a shared Wallet" />
-
-        <Button
-            android:id="@+id/button"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="Join!" />
     </LinearLayout>
 
 </FrameLayout>
