+++ demo-android/src/main/res/menu/bottom_navigation_menu.xml
@@ -23,7 +23,7 @@
 <!--        android:title="Debug"-->
 <!--        android:icon="@drawable/ic_bug_report_black_24dp" />-->
     <item
-        android:id="@+id/bitcoinFragment"
+        android:id="@+id/landingBitcoinFragment"
         android:title="@string/bitcoin"
         android:icon="@drawable/ic_device_hub_black_24dp" />
 </menu>
