+++ demo-android/src/main/res/navigation/nav_graph.xml
@@ -52,17 +52,16 @@
     <fragment
         android:id="@+id/bitcoinFragment"
         android:name="nl.tudelft.ipv8.android.demo.ui.bitcoin.BitcoinFragment"
-        android:label="fragment_bitcoin"
-        tools:layout="@layout/fragment_bitcoin" >
-        <action
-            android:id="@+id/action_bitcoinFragment_to_join_network"
-            app:destination="@id/join_network"
-            app:enterAnim="@anim/fragment_fade_enter"
-            app:exitAnim="@anim/fragment_close_exit" />
-    </fragment>
+        android:label="Bitcoin Overview"
+        tools:layout="@layout/fragment_bitcoin"></fragment>
     <fragment
-        android:id="@+id/join_network"
-        android:name="nl.tudelft.ipv8.android.demo_currency.join_network"
-        android:label="fragment_join_network"
+        android:id="@+id/joinNetworkFragment"
+        android:name="nl.tudelft.ipv8.android.demo.ui.bitcoin.JoinNetworkFragment"
+        android:label="Join Shared Wallet"
         tools:layout="@layout/fragment_join_network" />
+    <fragment
+        android:id="@+id/landingBitcoinFragment"
+        android:name="nl.tudelft.ipv8.android.demo.ui.bitcoin.LandingBitcoinFragment"
+        android:label="Bitcoin Overview"
+        tools:layout="@layout/fragment_landing_bitcoin" />
 </navigation>
